Support optional Providers
The Provider interface does require the dependency to be resolvable as returning null is explicitly allowed. Optional providers (those marked with @Opt or @Nullable) however are allowed to do so.

The InjectorStoreConsistencyPolicy will enforce that the dependency provisioned is present in the store unless it is optional. Providers (optional or not) can still be used to retrieve multiple instances, do lazy or optional retrieval, to break circular dependencies and to abstract scope.